% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_coverageplot.R
\name{add_sliding_windows}
\alias{add_sliding_windows}
\title{Add extended windows for coverage plot}
\usage{
add_sliding_windows(region, up_stream, down_stream)
}
\arguments{
\item{region}{Gene region for plotting}

\item{up_stream}{Upper range of the slide window}

\item{down_stream}{Lower range of the slide window}
}
\value{
Gene range for plot
}
\description{
This function is to add an extended window for ploting
}
