% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{add_tract}
\alias{add_tract}
\title{Add tracts to the circlize plot}
\usage{
add_tract(data_plot, group, colors = NULL)
}
\arguments{
\item{data_plot}{Data for circlize plot}

\item{group}{The group for showing on the new track}

\item{colors}{Color palette to color the group}
}
\value{
A new circlize track adding to the current circlize plot
}
\description{
This function allows users to add more tracks into the circlize plot
}
