% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{cell_order}
\alias{cell_order}
\title{Order the cells from each cluster}
\usage{
cell_order(dat)
}
\arguments{
\item{dat}{Data input.}
}
\value{
An vector with ordered cells
}
\description{
This function orders the cells from each cluster by giving a value from
1 to max
}
