% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_strip_background}
\alias{change_strip_background}
\title{A function to change the strip background color in ggplot}
\usage{
change_strip_background(ggplt_obj, type = "top", strip.color = NULL)
}
\arguments{
\item{ggplt_obj}{A ggplot object}

\item{type}{Strip on the "top" side only or "both" sides}

\item{strip.color}{A color vector}
}
\description{
A function to change the strip background color in ggplot
}
