% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{complex_dotplot_single}
\alias{complex_dotplot_single}
\title{Plot single gene across groups}
\usage{
complex_dotplot_single(
  seu_obj,
  feature,
  groupby,
  splitby = NULL,
  strip.color = NULL,
  scale.by = "radius"
)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{feature}{Gene name. Only one gene is allowed.}

\item{groupby}{The group to show on x axis. One of the column names in meta.data.}

\item{splitby}{The group to separate the gene expression. One of the column names in meta.data.}

\item{strip.color}{Colors for the strip background}

\item{scale.by}{Methods to scale the dot size. "radius" or "size"}
}
\value{
A ggplot object
}
\description{
This function can be used for plotting a single gene expression across
different groups in a study with complex group design. It uses the data
output from the DotPlot function in the Seurat package and uses ggplot2
to re-create a dotplot for better visualization. In the output graph, y axis
is the cluster name and x axis is the group ID.
}
