% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{complex_featureplot}
\alias{complex_featureplot}
\title{Plot gene expression on umap}
\usage{
complex_featureplot(
  seu_obj,
  features,
  group,
  select = NULL,
  cols = NULL,
  label.size = 12,
  order = F,
  strip.color = NULL,
  pt.size = 0.01
)
}
\arguments{
\item{seu_obj}{A complete Seurat object.}

\item{features}{Gene names to be plotted.}

\item{group}{The group to show on y axis. One of the column names in meta.data.}

\item{select}{Select the elements within the group to show.}

\item{cols}{Change the color legend.}

\item{label.size}{Change the label size.}

\item{strip.color}{Colors for the strip background.}

\item{pt.size}{Point size for each cell.}
}
\value{
A ggplot object
}
\description{
This function can be used for plotting a single gene or multiple genes expression across
different groups in a seurat featureplot format.
}
