% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{complex_heatmap_unique}
\alias{complex_heatmap_unique}
\title{Plot gene expression across groups using ComplexHeatmap}
\usage{
complex_heatmap_unique(
  seu_obj,
  celltype,
  group,
  gene_highlight = NULL,
  logfc = 0.5,
  return_marker = FALSE,
  col_fun = colorRamp2(c(-2, -1, 0, 1, 2), rev(c("#BF0080", "#CE6EAE", "#dddddd",
    "#6EAE6E", "#008000")))
)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{celltype}{Cell types selected for gene plot.}

\item{group}{The group to show on x axis. One of the column names in meta.data.}

\item{gene_highlight}{Gene names showing on the rows. Default: all genes}

\item{logfc}{Fold change to select the genes}

\item{return_marker}{If TRUE, a list of specific gene will be returned.}

\item{col_fun}{Heatmap color key.}
}
\value{
A ComplexHeatmap object or/and a gene list
}
\description{
This function is for identifying the group-specific genes in a selected celltype
and plot the expression of those genes in heatmap.
}
