% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{complex_upset_plot}
\alias{complex_upset_plot}
\title{UpSet plot visualize the number of unique and shared DEGs across group}
\usage{
complex_upset_plot(seu_obj, celltype, group, logfc = 0.5, min_size = 1)
}
\arguments{
\item{seu_obj}{A complete Seurat object.}

\item{celltype}{The cell type to analyze.}

\item{group}{Group factor in meta data.}

\item{logfc}{Log fold change to select the DEGs}

\item{min_size}{Minimal number of observations in an intersection for it to be included}
}
\value{
An UpSet plot
}
\description{
This function takes Seurat object as input and visualize the genes that
are unique to a particular group or shared by multiple group.
}
