% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{complex_vlnplot_single}
\alias{complex_vlnplot_single}
\title{Violin plot for a single gene across groups}
\usage{
complex_vlnplot_single(
  seu_obj,
  feature,
  celltypes = NULL,
  groups,
  add.dot = T,
  font.size = 14,
  pt.size = 0.1,
  splitby = NULL
)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{feature}{Gene name. Only one gene is allowed.}

\item{celltypes}{Cell types of interest. By default, all cell types are included.}

\item{groups}{Groups selected for plotting. Support multiple groups.}

\item{add.dot}{Whether or not to add points on the violins.}

\item{font.size}{Font size for the labels.}

\item{pt.size}{Point size for the data points on the violin}

\item{splitby}{Group to split the gene expression. Only works when length(groups)==1.}
}
\value{
A ggplot object
}
\description{
This function generates violin plot(s) to compare the expression of a single gene across
different groups or cell types. It is designed for visualizing a complicated scenario:
Gene expression on multiple cell types and multiple conditions.
}
