% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_geneid}
\alias{convert_geneid}
\title{A function to convert ensembl ID to gene name}
\usage{
convert_geneid(count, species)
}
\arguments{
\item{count}{Count data input}

\item{species}{value can be "mouse" or "human"}
}
\value{
A new count data with external gene names in the rows
}
\description{
This function is for replace the ensembl ID with actual gene names
}
