% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{creat_cellphonedb_file}
\alias{creat_cellphonedb_file}
\title{Generate the files required by CellPhoneDB}
\usage{
creat_cellphonedb_file(Seurat_obj, group)
}
\arguments{
\item{group}{The group selected for analysis}

\item{seu_obj}{A complete Seurat object}
}
\value{
Output two txt files for CellPhoneDB
}
\description{
This function can generate the files used as input into CellPhoneDB analysis.
Two files will be created: normalized count matrix with the designated group
and the metadata to descript each cell.
}
