% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_pyscenic_file}
\alias{create_pyscenic_file}
\title{Generate the files required by pySCENIC}
\usage{
create_pyscenic_file(seu_obj, celltypes, min.gene = 1)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{celltypes}{The cell types being selected for analysis}

\item{min.gene}{Cutoff to filter the low expression genes}
}
\value{
A loom file
}
\description{
This function can generate the file used as input into pySCENIC analysis.
A loom file will be created.
}
