% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_processing}
\alias{data_processing}
\title{A function to process the h5 file from CellBender and generate QC plots}
\usage{
data_processing(cellbender_h5, sampleID, out_dir, species, type)
}
\arguments{
\item{cellbender_h5}{The h5 file from CellBender output}

\item{sampleID}{ID given to the run}

\item{out_dir}{Output directory}

\item{species}{It can be "human" or "mouse"}

\item{type}{It can be "cell" or "nucleus". If your data is from snRNA-seq, use "type = nucleus". Otherwise, use "type = cell".}
}
\value{
QC plots and seurat objects before and after QC
}
\description{
This function works only with the h5 file output from CellBender. It takes the
count table, clusters the cells, removes the doublets (DoubletFinder), and
recluster the cells, and finally output a Seurat object
}
