% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_gene_count}
\alias{extract_gene_count}
\title{A function to extract gene counts for ploting}
\usage{
extract_gene_count(
  seu,
  features,
  cell.types = NULL,
  data.type = "data",
  meta.groups = NULL
)
}
\arguments{
\item{seu}{A finished Seurat Object with cell type annotation in the active.ident slot}

\item{features}{Gene names to extract expression data}

\item{cell.types}{The cell types to be inspected. By default, it will incorporate all cell types.}

\item{data.type}{The data slot to be accessed. By default, the "data" slot will be used.}

\item{meta.groups}{The colnames in the meta.data slot you want to include.}
}
\value{
A data frame with the requested info.
}
\description{
This function is a modified Seurat::FetchData function to extract gene
counts and the associated meta data for ploting. It returns a dataframe
with the requested information from the Seurat object.
}
