% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstup}
\alias{firstup}
\title{A function to change gene name into first letter capital}
\usage{
firstup(gene)
}
\arguments{
\item{gene}{Gene name}
}
\description{
The function is modified from this thread: https://stackoverflow.com/questions/18509527/first-letter-to-upper-case/18509816
}
