% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{get_segment}
\alias{get_segment}
\title{Create a segment for each element in a group}
\usage{
get_segment(dat, group)
}
\arguments{
\item{dat}{Data input.}

\item{group}{The group name}
}
\value{
An vector with ordered cells
}
\description{
This function creates a segment for each element within a group
}
