% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{mk_color_table}
\alias{mk_color_table}
\title{Create a dataframe for color mapping}
\usage{
mk_color_table(group)
}
\arguments{
\item{group}{Group to be assigned color}
}
\value{
A dataframe with colors assigned to groups
}
\description{
This function assigns a color for each value in a vector
}
