% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_coverageplot.R
\name{modified_coverageplot}
\alias{modified_coverageplot}
\title{Modified CoveragePlot}
\usage{
modified_coverageplot(
  sig_obj,
  gene,
  slide_windows_up,
  slide_windows_down,
  split_by
)
}
\arguments{
\item{sig_obj}{A complete Signac object}

\item{gene}{Gene name}

\item{slide_windows_up}{Upper range of the slide window}

\item{slide_windows_down}{Lower range of the slide window}

\item{split_by}{Group ID in metadata to split peaks}
}
\value{
A ggplot object
}
\description{
This function is a modified version of the CoveragePlot in Signac that allows gene name input and group splitting
}
