% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_seurat_plot.R
\name{modified_dimplot}
\alias{modified_dimplot}
\title{Modified DimPlot}
\usage{
modified_dimplot(
  seu_obj,
  colors = NULL,
  pt.size = 0.5,
  label.box = T,
  label.size = 6,
  title = "UMAP plot"
)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{colors}{Colors to label the clusters}

\item{pt.size}{Size of the data points}

\item{label.box}{Whether or not to label the cell type name}

\item{label.size}{Font size of the labels}

\item{title}{Main title of the plot}
}
\value{
A ggplot object
}
\description{
This function is a modified version of the DimPlot function in Seurat
}
