% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_seurat_plot.R
\name{modified_dotplot}
\alias{modified_dotplot}
\title{Modified DotPlot}
\usage{
modified_dotplot(seu_obj, features, col_palette = NULL, scale.by = "radius")
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{features}{A vector of gene names.}
}
\value{
A ggplot object
}
\description{
This function is a modified version of the DotPlot function in Seurat
}
