% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_gene_up}
\alias{order_gene_up}
\title{A function to order genes upregulated from the first to the last column}
\usage{
order_gene_up(df)
}
\arguments{
\item{df}{A data frames with genes in row and samples in column}
}
\description{
A function to order genes upregulated from the first to the last column
}
