% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_fraction.R
\name{plot_cell_fraction}
\alias{plot_cell_fraction}
\title{Plot cell fractions across groups}
\usage{
plot_cell_fraction(
  seu_obj,
  celltypes = NULL,
  groupby,
  show_replicate = FALSE,
  rep_colname = NULL,
  strip.color = NULL
)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{celltypes}{Cell types to be included in the plot. Default: all cell types.}

\item{groupby}{The group to show on x axis. One of the column names in meta.data.}

\item{show_replicate}{Whether or not to show the individual replicate on the graph. If TRUE, the replicate column name needs to specify in the argument rep_colname.}

\item{rep_colname}{The column name for biological replicates in the meta data.}

\item{strip.color}{Colors for the strip background}
}
\value{
A ggplot object
}
\description{
This function is to show the cell fraction changes across groups.
}
