% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{plot_circlize}
\alias{plot_circlize}
\title{Generate a circlize plot outside the tSNE/UMAP}
\usage{
plot_circlize(
  data_plot,
  do.label = T,
  contour.levels = c(0.2, 0.3),
  pt.size = 0.5,
  kde2d.n = 1000,
  contour.nlevels = 100,
  bg.color = "#F9F2E4"
)
}
\arguments{
\item{data_plot}{Data frame prepared by the prepare_circlize_data function}

\item{do.label}{Whether to label the clusters}

\item{contour.levels}{Which contour line to be drawn on the plot. Value: 0-1}

\item{pt.size}{Point size of the graph}

\item{kde2d.n}{Number of grid points in each direction. A kde2d parameter}

\item{contour.nlevels}{Total number of levels in contour}

\item{bg.color}{Canvas background color}
}
\value{
Return a circlize plot
}
\description{
This function generates a circlize plot outside the tSNE/UMAP
}
