% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_group.R
\name{plot_multigene_group}
\alias{plot_multigene_group}
\title{Plot multiple genes across groups}
\usage{
plot_multigene_group(seu_obj, features, splitby, strip.color = NULL)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{features}{A vector of gene names.}

\item{splitby}{Group ID Must be one of the column names in the meta.data slot of the Seurat object.}
}
\value{
A ggplot object
}
\description{
This function allows for visualization of multiple genes in multiple groups.
It takes the single gene expression data generated by PlotSingleGeneGroup,
concatenate all data, and produces a dotplot graph where the group ID are in
x axis, wrapped by cell types, genes are on the y axis.
}
