% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_extra.R
\name{plot_qpcr}
\alias{plot_qpcr}
\title{Plot the qPCR results}
\usage{
plot_qpcr(qPCR_file, metadata_file, ref_gene, ref_sample, file_name)
}
\arguments{
\item{qPCR_file}{Path to the Cq file from qPCR machine. Must be in csv format.}

\item{metadata_file}{Path to the metadata file to assign your samples into groups. Must be in csv format.}

\item{ref_gene}{The gene used for normalization. e.g. GAPDH.}

\item{ref_sample}{The sample used as reference sample. e.g. sample from the control group.}

\item{file_name}{The output file name}
}
\value{
There are three output files produced from this script. The csv file contains the average quantitative value (2^-ΔΔCt) for each sample (and each gene) after normalized by the reference gene (e.g. GAPDH) and the reference sample (e.g. sample from the control group). This file can be input into Graphpad Prism. The txt file includes all statistics from comparisons of any two given groups. If the run has two groups only, Welch's t test will be performed. Otherwise, one-way ANOVA with post-hoc Tukey's test will be performed. Finally, the tiff file is a boxplot graph to visualize the gene expression across groups.
}
\description{
This function is to process the Cq file generated by the qPCR machine
in our lab.
}
