% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_group.R
\name{plot_singlegene_group}
\alias{plot_singlegene_group}
\title{Plot single gene across groups}
\usage{
plot_singlegene_group(seu_obj, feature, splitby)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{feature}{Gene name. Only one gene is allowed.}

\item{splitby}{One of the column names in the meta.data slot of the Seurat object.}
}
\value{
A ggplot object
}
\description{
This function can be used for plotting a single gene expression across
different groups in a study with complex group design. It uses the data
output from the DotPlot function in the Seurat package and uses ggplot2
to re-create a dotplot for better visualization. In the output graph, y axis
is the cluster name and x axis is the group ID.
}
