% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{plot_upset}
\alias{plot_upset}
\title{UpSet plot for the cell markers}
\usage{
plot_upset(
  deg_table,
  celltypes = NULL,
  nintersects = NULL,
  textscale = 1.5,
  pointsize = 2,
  alpha = 0.3
)
}
\arguments{
\item{deg_table}{The dataframe from Seurat::FindAllMarkers function.}

\item{celltypes}{The cell types of interest}

\item{nintersects}{The number of intersections showing in the main bar}

\item{textscale}{Scale of the text font size}

\item{pointsize}{The size of the dots in the matrix}

\item{alpha}{The transparency of the bars}
}
\value{
An UpSet plot
}
\description{
This function takes the marker gene table as input and visualize the genes that
are unique to a particular cell type or shared by multiple cell types.
}
