% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{prepare_circlize_data}
\alias{prepare_circlize_data}
\title{Prepare circlize data for plotting}
\usage{
prepare_circlize_data(seu_obj, scale = 0.8)
}
\arguments{
\item{seu_obj}{Seurat object}

\item{scale}{Scale factor to zoom in our zoom out the tSNE/UMAP proportionally}
}
\value{
A data frame for plotting
}
\description{
This function creates a segment for each element within a group
}
