% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_correlation}
\alias{run_correlation}
\title{Compute and plot correlations between two datasets}
\usage{
run_correlation(
  data1,
  data2,
  ngenes = 2000,
  method.use = "pearson",
  color.use = NULL
)
}
\arguments{
\item{data1}{The first data frame with genes in rows and samples in columns}

\item{data2}{The second data frame with genes in rows and samples in columns}

\item{ngenes}{Number of top variable genes used for the computation}

\item{method.use}{Default is "pearson". Other methods include kendall" or "spearman". See ?cor.}

\item{color.use}{Color palette for the heatmap plot.}
}
\value{
A heatmap plot
}
\description{
This function can compute the correlations on the samples from two count matrix.
It will use Pearson method as default.
}
