% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{transform_coordinates}
\alias{transform_coordinates}
\title{Convert coordinates}
\usage{
transform_coordinates(coord_data, zoom)
}
\arguments{
\item{coord_data}{Cartesian coordinates from tSNE, UMAP, etc.}

\item{zoom}{value from c(0,1) to adjust the coordinates.}
}
\value{
A matrix with polar coordinates
}
\description{
This function converts the Cartesian coordinates to Polar coordinates.
Input data can be the coordinates from tSNE or UMAP. It outputs a matrix with
polar coordinates.
}
