% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{complex_dotplot_multiple}
\alias{complex_dotplot_multiple}
\title{Plot multiple genes across groups}
\usage{
complex_dotplot_multiple(seu_obj, features, groupby, strip.color = NULL)
}
\arguments{
\item{seu_obj}{A complete Seurat object}

\item{features}{A vector of gene names.}

\item{groupby}{Group ID Must be one of the column names in the meta.data slot of the Seurat object.}

\item{strip.color}{Colors for the strip background}
}
\value{
A ggplot object
}
\description{
This function allows for visualization of multiple genes in multiple groups.
It takes the single gene expression data generated by PlotSingleGeneGroup,
concatenate all data, and produces a dotplot graph where the group ID are in
x axis, wrapped by cell types, genes are on the y axis.
}
