% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{complex_dotplot_single}
\alias{complex_dotplot_single}
\title{Plot single gene across groups}
\usage{
complex_dotplot_single(
  seu_obj,
  feature,
  celltypes = NULL,
  groups,
  splitby = NULL,
  color.palette = NULL,
  font.size = 12,
  strip.color = NULL,
  do.scale = T,
  scale.by = "radius"
)
}
\arguments{
\item{seu_obj}{A complete Seurat object.}

\item{feature}{Gene name. Only one gene is allowed.}

\item{celltypes}{Cell types to be included in the dot plot. Default: all cell types.}

\item{groups}{The group to show on x axis. One of the column names in meta.data.}

\item{splitby}{The group to separate the gene expression. One of the column names in meta.data.}

\item{color.palette}{Color for gene expression.}

\item{font.size}{Font size for the labels.}

\item{strip.color}{Colors for the strip background.}

\item{do.scale}{Whether or not to scale the dot when percentage expression of the gene is less than 20.}

\item{scale.by}{Methods to scale the dot size. "radius" or "size".}
}
\value{
A ggplot object
}
\description{
This function can be used for plotting a single gene expression across
different groups in a study with complex group design.
}
