% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{mk_marker_ct}
\alias{mk_marker_ct}
\title{Make count matrix for the selected markers}
\usage{
mk_marker_ct(obj, features)
}
\arguments{
\item{obj}{SeuratObject}

\item{features}{Selected marker genes}
}
\value{
A dataframe with cells labeled by marker genes
}
\description{
This function labels the cells based their expression levels of the selected
marker genes.
}
